Possibly last update.
So the goal here was to
A: generate data on the CPU (in rust)
B: repeatedly preform calculations on that data using a GPU
C: send more data that can affect the calculations, that can be updated with each iteration.
D: render the output of the compute shader using vertex and fragment shaders, also on the GPU.
All of those were met, and if I ever need to do something like this again, I can.
Signed-off-by: Owen McKenna <owenpmckenna@outlook.com>